# platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDeskViewer RustDesk Pro API
  version: 1.0.0
  description: |
    RustDesk Pro-compatible API for OpenDeskViewer management platform.
    This API is consumed by the unmodified RustDesk client.
    
    All endpoints use snake_case and return `{"error": "..."}` on failure.
    Pagination uses `?current=&pageSize=` with response `{total, data}`.
    Authentication: `Authorization: Bearer <token>` for authenticated endpoints.
servers:
  - url: http://localhost:8000
paths:
...
  /api/peers:
    get:
      summary: Get peers list
      description: Returns all peers in the system (Admin only).
      operationId: getPeers
      tags:
        - Peer
      security:
        - bearerAuth: []
      parameters:
        - name: current
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Peers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Peer'
                  total:
                    type: integer
                required:
                  - data
                  - total

  # ========== Heartbeat & Sysinfo ==========
...
  /api/v1/device-groups:
    get:
      summary: List device groups
      operationId: listDeviceGroups
      tags:
        - Portal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Current'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A page of device groups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PortalDeviceGroup'
    post:
      summary: Create a device group
      operationId: createDeviceGroup
      tags:
        - Portal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                parent_id:
                  type: string
              required:
                - name
      responses:
        '201':
          description: The created device group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalDeviceGroup'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Current:
      name: current
      in: query
      description: 1-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Rows per page. Capped at 500 by the server.
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 10
...
  schemas:
...
    Peer:
      type: object
      properties:
        id:
          type: string
          description: Device ID (rustdesk_id)
        uuid:
          type: string
        username:
          type: string
        hostname:
          type: string
        platform:
          type: string
          enum: [Windows, Linux, macOS, Android, iOS]
        tags:
          type: string
        online:
          type: boolean
        status:
          type: string
          enum: [online, offline]
        connect_status:
          type: string
        login_name:
          type: string
        last_connected_at:
          type: string
          format: date-time
        device_group_id:
          type: string
        customer_id:
          type: string
        location_id:
          type: string
      required:
        - id
        - uuid
        - username
        - hostname
        - platform
...
    PageEnvelope:
      description: >
        The list envelope. total is the count matching the filters, not the
        length of data, so a client paging with `while current * pageSize <
        total` terminates correctly.
      type: object
      properties:
        total:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PortalDeviceGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parent_id:
          type: string
          format: uuid
        member_count:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...